# backend/openapischema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Document Analysis API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/collections/": {
      "get": {
        "tags": ["Collections"],
        "summary": "Retrieve all collections",
        "description": "Fetches a list of all collections available in the database, including their associated metadata and document counts.",
        "operationId": "read_collections_api_v1_collections__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CollectionInDB"
                  },
                  "type": "array",
                  "title": "Response Read Collections Api V1 Collections  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Collections"],
        "summary": "Create a new collection",
        "description": "Creates a new collection with the provided name, description, and tags. The collection's ID, creation timestamp, and document count are returned upon successful creation.",
        "operationId": "create_collection_api_v1_collections__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections/{collection_id}": {
      "get": {
        "tags": ["Collections"],
        "summary": "Retrieve a single collection by ID",
        "description": "Fetches a single collection's details using its unique identifier. Returns a 404 error if the collection is not found.",
        "operationId": "read_collection_api_v1_collections__collection_id__get",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Collections"],
        "summary": "Update an existing collection",
        "description": "Updates an existing collection's details using its unique identifier. Allows for partial updates (e.g., changing only the name or description). Returns a 404 error if the collection is not found.",
        "operationId": "update_collection_api_v1_collections__collection_id__patch",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Collections"],
        "summary": "Delete a collection",
        "description": "Deletes a collection and all of its associated documents using its unique identifier. Returns a 404 error if the collection is not found.",
        "operationId": "delete_collection_api_v1_collections__collection_id__delete",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/recommendations/collections/{collection_id}": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "Retrieve all recommendations for a collection",
        "description": "Fetches a list of all recommendation records associated with a specific collection. The response includes basic recommendation details but excludes the full content of associated items for brevity.",
        "operationId": "get_all_recommendations_for_collection_api_v1_recommendations_collections__collection_id__get",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the collection to retrieve recommendations for.",
              "title": "Collection Id"
            },
            "description": "The ID of the collection to retrieve recommendations for."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get All Recommendations For Collection Api V1 Recommendations Collections  Collection Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollectionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the collection is required on creation."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "A brief description of the collection's purpose."
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "A list of tags for categorization."
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "CollectionCreate",
        "description": "Pydantic model for creating a new collection.\n\nInherits from CollectionBase and makes the 'name' field required."
      },
      "CollectionInDB": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "The name of the collection."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "A brief description of the collection's purpose."
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "A list of tags for categorization."
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "The unique identifier for the collection."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat",
            "description": "The timestamp when the collection was created."
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat",
            "description": "The timestamp of the last update."
          },
          "total_docs": {
            "type": "integer",
            "title": "Total Docs",
            "description": "The total number of documents in the collection.",
            "default": 0
          },
          "latestInsightId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestinsightid",
            "description": "ID of the latest insight generated for this collection."
          },
          "latestPodcastId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestpodcastid",
            "description": "ID of the latest podcast generated for this collection."
          }
        },
        "type": "object",
        "required": ["id", "createdAt"],
        "title": "CollectionInDB",
        "description": "Pydantic model representing a collection as it is stored in the database.\n\nThis model is used for API responses and includes database-generated fields."
      },
      "CollectionUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "The name of the collection."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "A brief description of the collection's purpose."
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "A list of tags for categorization."
          }
        },
        "type": "object",
        "title": "CollectionUpdate",
        "description": "Pydantic model for updating an existing collection.\n\nAll fields are optional to allow for partial updates."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}